Sort grammars by name and scope name
Previously a spec was flaky since it assumed an order in the grammar table
which could be different each time it runs since the grammars are loaded and
registered asynchronously which affected the order in the table.

Now the grammars are sorted by name and scope name to ensure a consistent
and expected ordering.