Normalize order of displayed packages
This is not a major issue, just a minor annoyance that I've noticed over the past few weeks. The order in which installed packages are displayed seems unnatural to me for the following reasons:
* `dev-packages` generally make up the smallest sub-section of installed packages, they should therefore be displayed first
* `dev-packages` are also adjusted more often than `community` or `core` packages, by people trying to tweak them before a release
* `dev-packages` take precedence over `community-packages`, so they should be listed before, not after the `community-packages`
* `core-packages` are almost never going to change, and will therefore be needed least often and should be displayed last

Like I said, not a big deal, it's just another one of those minor details that, if adjusted, will make the user experience just that much better!