:bug: Wait for updated package to load and then try to activate
This is a fix for issue #532 - Package cards in Updates panel don't change state when updates are done installing.

Before this commit, updating a package that had activation commands would cause the package card to display the loading state forever.

The cause of the issue was that the update process was waiting for the updated package to activate. This was a problem because calling `activatePackage` will only immediately activate a package that doesn't have any activationCommands.

The proposed fix is to wait for the package to load and then attempt to activate the package without waiting.