Fix the problem that tabLength of language scope jumps back to global setting automatically.
When I set tabLength to 4 in global editor setting, and set it to 2 in
language scope setting (such as coffeescript or scala), the one in
language setting immediately jumps back to 4. Even when I set it in config
file, it still jumps back. It turned out the settings-view does not look
at global settings when resetting to default. So I did a small patch.