Implement custom renderer for config options
This commit provides handling for a new field on config properties
called `renderer`. If this option is provided, it's assumed to be a
class which is an Atom view. This class will be instantiated and
attached to the settings panel in place of the generic settings view,
which allows for implementation of more advanced configuration
interfaces.

Closes #90.